+++ templates/default/payments/manage.js.tpl
