+++ templates/default/payments/manage.tpl
@@ -29,7 +29,7 @@
 			</div>        		
         {else}
         	<table id="manageGrid" style="display:none"></table>
-        	{include file='../modules/payments/manage.js.php' get=$smarty.get}
+        	{include file='../templates/default/payments/manage.js.tpl' get=$smarty.get}
         {/if}
 
 	{elseif $smarty.get.c_id }
@@ -41,7 +41,7 @@
 			</div>        		
         {else}
     	    <table id="manageGrid" style="display:none"></table>
-        	{include file='../modules/payments/manage.js.php' get=$smarty.get}
+        	{include file='../templates/default/payments/manage.js.php' get=$smarty.get}
         {/if}
 
 	{else}
@@ -53,7 +53,7 @@
 			</div>        		
         {else}
         	<table id="manageGrid" style="display:none"></table>
-        	{include file='../modules/payments/manage.js.php' get=$smarty.get}
+        	{include file='../templates/default/payments/manage.js.php' get=$smarty.get}
         {/if}
 
 	{/if}
